HSEARCH-4522 Rename integrationtest/jdk/java-modules directory and packages
To match the artifact name.
