--- integrationtest/jdk/java-modules/src/main/resources/hibernate.properties
@@ -1,15 +1,0 @@
-# Hibernate ORM properties:
-## Connection info: see integration-test parent POM
-hibernate.hbm2ddl.auto = create-drop
-hibernate.show_sql = true
-hibernate.format_sql = true
-hibernate.max_fetch_depth = 5
-# We can't use classes from the hibernate-testing module unless we add an explicit dependency to that module.
-#hibernate.cache.region_prefix hibernate.test
-#hibernate.cache.region.factory_class = org.hibernate.testing.cache.CachingRegionFactory
-
-# Hibernate Search properties:
-## Connection info: see POM
-hibernate.search.automatic_indexing.synchronization.strategy = sync
-hibernate.search.backend.log.json_pretty_printing = true
-hibernate.search.backend.analysis.configurer = org.hibernate.search.integrationtest.java.module.config.MyElasticsearchAnalysisConfigurer
