+++ integrationtest/pom.xml
@@ -41,7 +41,7 @@
                 </property>
             </activation>
             <modules>
-                <module>jdk/java-modules</module>
+                <module>java/modules</module>
             </modules>
         </profile>
     </profiles>
