--- orm6/integrationtest/jdk/java-modules/pom.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.hibernate.search</groupId>
-        <artifactId>hibernate-search-integrationtest-orm6</artifactId>
-        <version>6.2.0-SNAPSHOT</version>
-        <relativePath>../../</relativePath>
-    </parent>
-    <artifactId>hibernate-search-integrationtest-java-modules-orm6</artifactId>
-    <packaging>jar</packaging>
-
-    <name>Hibernate Search ITs - JDK - Java modules - ORM6</name>
-    <description>Hibernate Search integration tests for (JDK 11+) Java modules - ORM6 version</description>
-
-    <properties>
-        <transform.original.pathFromRoot>integrationtest/jdk/java-modules</transform.original.pathFromRoot>
-
-        <test.elasticsearch.run.skip>${test.elasticsearch.connection.uris.defined}</test.elasticsearch.run.skip>
-        <test.database.run.skip>false</test.database.run.skip>
-
-        <maven.compiler.release>11</maven.compiler.release>
-        <!-- Override the default from the parent POM: here we really do want to use the modulepath -->
-        <failsafe.useModulePath>true</failsafe.useModulePath>
-
-        <surefire.jvm.args.module>
-            ${test.elasticsearch.connection.jvm.args}
-        </surefire.jvm.args.module>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.hibernate.search</groupId>
-            <artifactId>hibernate-search-mapper-orm-orm6</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.hibernate.search</groupId>
-            <artifactId>hibernate-search-backend-elasticsearch</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>${jdbc.driver.groupId}</groupId>
-            <artifactId>${jdbc.driver.artifactId}</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-antrun-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>it</id>
-                        <goals>
-                            <goal>integration-test</goal>
-                            <goal>verify</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>io.fabric8</groupId>
-                <artifactId>docker-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
